pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = ""
authors = ["Ivan <vandal12313@gmail.com>"]
# readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
asyncpg = ">=0.30.0,<0.31.0"
alembic = ">=1.16.5,<2.0.0"
uvicorn = ">=0.35.0,<0.36.0"
python-dotenv = ">=0.21,<2.0"
pydantic = { extras = ["email"], version = ">=2.11.9,<3.0.0" }
python-jose = { extras = ["cryptography"], version = ">=3.5.0,<4.0.0" }
passlib = { extras = ["bcrypt"], version = ">=1.7.4,<2.0.0" }
libgravatar = ">=1.0.4,<2.0.0"
python-multipart = ">=0.0.20,<0.0.21"
fastapi-mail = ">=1.5.0,<2.0.0"
slowapi = ">=0.1.9,<0.2.0"
cloudinary = ">=1.44.1,<2.0.0"
fastapi = "^0.116.1"
psycopg2-binary = "^2.9.10"
greenlet = "^3.2.4"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"